Gave quirium cascades an entity class of their very own.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@4239 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/WormholeEntity.m
@@ -459,7 +459,7 @@ static void DrawWormholeCorona(GLfloat inner_radius, GLfloat outer_radius, int s
 	if (translucent)
 	{
 		// for now, a simple copy of the energy bomb draw routine
-		double srzd = sqrt(zero_distance);
+		float srzd = sqrtf(zero_distance);
 		
 		GLfloat	color_fv[4] = { 0.0, 0.0, 1.0, 0.25};
